Removed unessesary old forms of Integer type: php dir

+++ php/php.editor/test/unit/src/org/netbeans/modules/php/editor/indent/PHPFormatterQATest.java
@@ -548,7 +548,7 @@ public class PHPFormatterQATest extends PHPFormatterTestBase {
             String option = entry.getKey();
             Object value = entry.getValue();
             if (value instanceof Integer) {
-                prefs.putInt(option, ((Integer)value).intValue());
+                prefs.putInt(option, (Integer) value);
             }
             else if (value instanceof String) {
                 prefs.put(option, (String)value);
