+++ php/php.editor/test/unit/src/org/netbeans/modules/php/editor/indent/PHPNewLineIndenterTest.java
@@ -1274,7 +1274,7 @@ public class PHPNewLineIndenterTest extends PHPTestBase {
             String option = entry.getKey();
             Object value = entry.getValue();
             if (value instanceof Integer) {
-                prefs.putInt(option, ((Integer) value).intValue());
+                prefs.putInt(option, (Integer) value);
             } else if (value instanceof String) {
                 prefs.put(option, (String) value);
             } else if (value instanceof Boolean) {
